tests: doctest addition
* Adds doctests to ensure documentation examples works.
