+++ .travis.yml
@@ -17,7 +17,7 @@ python:
   - "3.4"
 
 install:
-  - pip install coveralls pep257 Sphinx
+  - pip install coveralls pep257 Sphinx httpretty
   - pip install pytest pytest-pep8 pytest-cov pytest-cache
   - pip install -e .
 
@@ -25,6 +25,7 @@ script:
   - pep257 cernservicexml
   - "sphinx-build -qnNW docs docs/_build/html"
   - python setup.py test
+  - "sphinx-build -qnNW -b doctest docs docs/_build/doctest"
 
 after_success:
   - coveralls
